+++ documentation/ChangeLog.md
@@ -1,6 +1,7 @@
 ## release ChangeLog
 
 ### (NextVersion)
+ * Added insertion of function parameters for function completion proposals.
  * Enabled gocode `propose-builtins` option, which isn't enabled by default.
  * Reworked the icon and label of completion proposals and fixed some related issues. Also, now the icons are consistent with the Outline icons. 
  * Fixed: Progress dialog never showing up when "Ok" pressed in LANG_NAME project properties page.